#544 Move new validation to src/components/plugins dir
