+++ client/src/components/plugins/v-model.js
@@ -1,0 +1,25 @@
+import { computed, toRefs } from 'vue';
+
+export const modelValueMixin = {
+  props: {
+    modelValue: {},
+  },
+  emits: ['update:modelValue'],
+};
+
+export const useModelValue = (props, context) => {
+  const { modelValue } = toRefs(props);
+
+  const vModel = computed({
+    get () {
+      return modelValue.value;
+    },
+    set (value) {
+      context.emit('update:modelValue', value);
+    },
+  });
+
+  return {
+    vModel,
+  };
+};
